requirements.txt
click==8.0.1
dataclasses_json==0.5.6
ImageHash==4.2.1
Pillow==8.4.0
typing_extensions==4.0.1
